# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Escolar",
        "description": "API para sistema de gerenciamento escolar multi-tenant. Permite cadastro de usuários, alunos, turmas, diários, eventos, entre outros. O contexto do tenant é definido pelo header 'x-tenant-id'.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "alexseyf66@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Servidor Local"
        },
        {
            "url": "https://elo-api-alexseyfs-projects.vercel.app",
            "description": "Servidor de Produção"
        }
    ],
...
    "paths": {
...
        "/turmas": {
            "get": {
                "tags": ["Turmas"],
                "summary": "Listar turmas",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de turmas" } }
            },
            "post": {
                "tags": ["Turmas"],
                "summary": "Criar turma",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": { "application/json": { "schema": { "type": "object", "properties": { "nome": { "type": "string" } } } } }
                },
                "responses": { "201": { "description": "Turma criada" } }
            }
        },
        "/turmas/totalAlunosTurma": {
            "get": {
                "tags": ["Turmas"],
                "summary": "Total de alunos por turma",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Totais obtidos" } }
            }
        },
        "/turmas/{usuarioId}/professor": {
            "post": {
                "tags": ["Turmas"],
                "summary": "Adicionar professor à turma",
                "parameters": [
                    { "name": "usuarioId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": { "application/json": { "schema": { "type": "object", "properties": { "turmaId": { "type": "integer" } }, "required": ["turmaId"] } } }
                },
                "responses": { "201": { "description": "Professor vinculado" } }
            }
        },
        "/turmas/{turmaId}/alunos": {
            "get": {
                "tags": ["Turmas"],
                "summary": "Listar alunos da turma",
                "parameters": [
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de alunos" } }
            }
        },
        "/professores/{professorId}/turmas": {
            "get": {
                "tags": ["Professores"],
                "summary": "Turmas de um professor",
                "parameters": [
                    { "name": "professorId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de turmas" } }
            }
        },
        "/professores/professor-turma": {
            "get": {
                "tags": ["Professores"],
                "summary": "Lista professores e suas turmas",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista obtida" } }
            }
        },
        "/professores/{professorId}/turma/{turmaId}": {
            "delete": {
                "tags": ["Professores"],
                "summary": "Desvincular professor de turma",
                "parameters": [
                    { "name": "professorId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Desvinculado com sucesso" } }
            }
        },
        "/responsaveis/{responsavelId}/alunos": {
            "get": {
                "tags": ["Responsaveis"],
                "summary": "Alunos de um responsável",
                "parameters": [
                    { "name": "responsavelId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de alunos" } }
            }
        },
        "/responsaveis/meus-alunos": {
            "get": {
                "tags": ["Responsaveis"],
                "summary": "Alunos do responsável logado",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de alunos" } }
            }
        },
...
        "/eventos/turma/{turmaId}": {
            "get": {
                "tags": ["Eventos"],
                "summary": "Listar eventos da turma",
                "parameters": [
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de eventos" } }
            }
        },
...
        "/objetivos": {
            "get": {
                "tags": ["Objetivos"],
                "summary": "Listar objetivos",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista" } }
            },
            "post": {
                "tags": ["Objetivos"],
                "summary": "Criar objetivo",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
                "responses": { "201": { "description": "Criado" } }
            }
        },
        "/objetivos/turma/{turmaId}": {
            "get": {
                "tags": ["Objetivos"],
                "summary": "Objetivos da turma",
                "parameters": [
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista" } }
            }
        },
        "/objetivos/grupo-campo": {
            "get": {
                "tags": ["Objetivos"],
                "summary": "Buscar por grupo e campo",
                "parameters": [
                    { "name": "grupoId", "in": "query", "required": true, "schema": { "type": "integer" } },
                    { "name": "campoId", "in": "query", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista" } }
            }
        },
...
    },
    "components": {
        "parameters": {
            "TenantId": {
                "name": "x-tenant-id",
                "in": "header",
                "description": "Slug ou ID da escola (Tenant)",
                "required": true,
                "schema": { "type": "string" }
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}